rns_management_backend/requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
beautifulsoup4==4.12.3
billiard==4.2.0
celery==5.3.6
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==4.2.11
djangorestframework==3.15.1
idna==3.6
kombu==5.3.5
prompt-toolkit==3.0.43
python-dateutil==2.9.0.post0
pytz==2024.1
redis==5.0.3
requests==2.31.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
typing-extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
psycopg2-binary==2.9.9
python-dotenv==1.0.1
django-cors-headers==3.7.0
